+++ tests/test-popper.js
@@ -45,6 +45,18 @@ describe('Popper.js', function() {
         pop.destroy();
     });
 
+    it('inits a bottom-start popper', function() {
+        var reference = appendNewRef(1);
+        reference.style.marginLeft = '200px';
+        var popper    = appendNewPopper(2);
+
+        var pop = new TestPopper(reference, popper, {placement: 'bottom-start'});
+
+        expect(popper.getBoundingClientRect().left).toBeApprox(reference.getBoundingClientRect().left);
+
+        pop.destroy();
+    });
+
     it('inits a right popper', function() {
         var reference = appendNewRef(1);
         var popper    = appendNewPopper(2);
@@ -213,7 +225,7 @@ describe('Popper.js', function() {
             // force redraw
             window.dispatchEvent(new Event('resize'));
 
-            var local = 729;
+            var local = 730;
             var ci = 739;
             expect([local, ci]).toContain(popper.getBoundingClientRect().top);
             expect(popper.getBoundingClientRect().left).toBeApprox(5);
@@ -240,6 +252,7 @@ describe('Popper.js', function() {
         new TestPopper(reference).onCreate(function(instance) {
             expect(document.querySelectorAll('.popper').length).toBe(1);
             document.body.removeChild(instance._popper);
+            instance.destroy();
             done();
         });
     });
@@ -252,6 +265,7 @@ describe('Popper.js', function() {
         }).onCreate(function(instance) {
             expect(instance._popper.innerText).toBe('something');
             document.body.removeChild(instance._popper);
+            instance.destroy();
             done();
         });
     });
@@ -263,7 +277,11 @@ describe('Popper.js', function() {
             modifiersIgnored: ['applyStyle']
         }).onUpdate(function(data) {
             expect(data.offsets.popper.top).toBeApprox(46);
+            document.body.removeChild(data.instance._popper);
+            data.instance.destroy();
             done();
+        }).onCreate(function(instance) {
+            instance.update();
         });
     });
 
@@ -274,6 +292,54 @@ describe('Popper.js', function() {
             expect(instance._popper).toBeDefined();
             expect(instance._popper.innerText).toBe('');
             document.body.removeChild(instance._popper);
+            instance.destroy();
+            done();
+        });
+    });
+
+    it('creates a popper with an empty form as parent, then auto remove it on destroy', function(done) {
+        var form = document.createElement('form');
+        var reference = appendNewRef(1, 'ref', form);
+        jasmineWrapper.appendChild(form);
+
+        new TestPopper(reference, { parent: form, content: 'test'}, { removeOnDestroy: true }).onCreate(function(instance) {
+            expect(instance._popper).toBeDefined();
+            expect(instance._popper.innerText).toBe('test');
+            instance.destroy();
+            expect(document.contains(instance._popper)).toBeFalsy();
+            done();
+        });
+    });
+
+    it('creates a popper with a not empty form as parent, then auto remove it on destroy', function(done) {
+        var form = document.createElement('form');
+        var input = document.createElement('input');
+        form.appendChild(input);
+        var reference = appendNewRef(1, 'ref', form);
+        jasmineWrapper.appendChild(form);
+
+        new TestPopper(reference, { parent: form, content: 'test'}, { removeOnDestroy: true }).onCreate(function(instance) {
+            expect(instance._popper).toBeDefined();
+            expect(instance._popper.innerText).toBe('test');
+            instance.destroy();
+            expect(document.contains(instance._popper)).toBeFalsy();
+            done();
+        });
+    });
+
+    it('creates a popper and make sure it\'s position is correct on init', function(done) {
+        var reference = appendNewRef(1);
+
+        new TestPopper(
+            reference,
+            { content: 'popper', classNames: ['none'] },
+            { placement: 'right', removeOnDestroy: true }
+        ).onCreate(function(instance) {
+            instance.update();
+            var local = 87;
+            var ci = 105;
+            expect([local, ci]).toContain(instance._popper.getBoundingClientRect().left);
+            instance.destroy();
             done();
         });
     });